+++ src/usr/install.rs
@@ -20,6 +20,7 @@ pub fn copy_files(verbose: bool) {
     copy_file("/bin/clear", include_bytes!("../../dsk/bin/clear"), verbose);
     copy_file("/bin/halt", include_bytes!("../../dsk/bin/halt"), verbose);
     copy_file("/bin/hello", include_bytes!("../../dsk/bin/hello"), verbose);
+    copy_file("/bin/print", include_bytes!("../../dsk/bin/print"), verbose);
     copy_file("/bin/reboot", include_bytes!("../../dsk/bin/reboot"), verbose);
 
     create_dir("/dev/clk", verbose); // Clocks
